requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0
Flask>=2.0.0
opencv-python
tensorflow>=2.10.0